requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.6.0
google-generativeai>=0.5.0
chromadb>=0.4.24
sentence-transformers>=2.6.0
streamlit>=1.32.0
requests>=2.31.0